Replaced Micrometer 1.3.3 with 1.5.1
Replaced sla method with serviceLevelObjectives and used executor prefix

Some further cleanup for the new micrometer lib

+++ com.openexchange.metrics.micrometer/src/com/openexchange/metrics/micrometer/internal/filter/DistributionMaximumMicrometerFilterPerformer.java
@@ -92,6 +92,6 @@ public class DistributionMaximumMicrometerFilterPerformer extends AbstractMicrom
         }
 
         long nanos = TimeUnit.MILLISECONDS.toNanos(TimeSpanParser.parseTimespanToPrimitive(value));
-        return DistributionStatisticConfig.builder().maximumExpectedValue(Long.valueOf(nanos)).build().merge(config);
+        return DistributionStatisticConfig.builder().maximumExpectedValue(Double.valueOf(nanos)).build().merge(config);
     }
 }
